# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Simple AI Gateway",
        "description": "LLM API \u7f51\u5173 \u2014 \u517c\u5bb9 OpenAI API \u534f\u8bae\uff0c\u652f\u6301\u591a\u4f9b\u5e94\u5546\u8def\u7531\u3001\u7194\u65ad\u3001\u9650\u6d41\u3001\u914d\u989d\u3001\u52a8\u6001\u914d\u7f6e\u7ba1\u7406\u4e0e Webhook \u4e8b\u4ef6\u901a\u77e5\u3002\n\n## API \u5206\u7ec4\n- **\u4e1a\u52a1\u7aef\u70b9**: `/v1/chat/completions`, `/v1/models`\n- **\u8ba4\u8bc1**: `/auth/*`\n- **\u7ba1\u7406**: `/admin/*`\n- **\u5185\u90e8**: `/internal/*`\n- **\u7cfb\u7edf**: `/healthz`\n\n## \u8ba4\u8bc1\u65b9\u5f0f\nBearer token: JWT (access token) \u6216 API Key",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        },
        {
            "ApiKeyAuth": []
        }
    ],
...
    "paths": {
...
        "/admin/users/{username}": {
            "put": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "putUserRole",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserView"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/users/{username}/limits": {
            "put": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "updateUserLimits",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserLimitsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{username}/api-keys/{keyId}/toggle": {
            "put": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "toggleUserApiKey",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApiKeyToggleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/users/{username}/allowed-models": {
            "put": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "updateUserAllowedModels",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserAllowedModelsUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserView"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/users": {
            "get": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "users",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "createUser",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{username}/reset-password": {
            "post": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "resetUserPassword",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResetPasswordResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{username}/api-keys": {
            "get": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "listUserApiKeys",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApiKeyView"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "createUserApiKey",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateApiKeyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyView"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/users/{username}/api-keys/{keyId}/rotate": {
            "post": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "rotateUserApiKey",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyRecord"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/users/{username}/api-keys/{keyId}": {
            "delete": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "deleteUserApiKey",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "admin-user-controller"
                ],
                "operationId": "updateUserApiKey",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "keyId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchApiKeyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateUserRequest": {
                "type": "object",
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "frozen": {
                        "type": "boolean"
                    }
                }
            },
...
            "UserView": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "apiKeyMasked": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "frozen": {
                        "type": "boolean"
                    },
                    "frozenAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "limits": {
                        "$ref": "#/components/schemas/UserLimits"
                    },
                    "allowedModels": {
                        "uniqueItems": true,
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "UpdateUserLimitsRequest": {
                "type": "object",
                "properties": {
                    "dailyTokens": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "monthlyTokens": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tokensPerMinute": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "maxTokens": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "dailyCost": {
                        "type": "number"
                    },
                    "monthlyCost": {
                        "type": "number"
                    }
                }
            },
            "ApiKeyToggleRequest": {
                "type": "object",
                "properties": {
                    "enabled": {
                        "type": "boolean",
                        "description": "Target enabled status",
                        "example": true
                    }
                },
                "description": "Request body for enabling or disabling a user API key"
            },
            "UserAllowedModelsUpdateRequest": {
                "type": "object",
                "properties": {
                    "allowedModels": {
                        "type": "array",
                        "description": "Models this user is allowed to call",
                        "example": [
                            "gpt-4o-mini",
                            "claude-3-5-sonnet"
                        ],
                        "items": {
                            "type": "string",
                            "description": "Models this user is allowed to call",
                            "example": "[\"gpt-4o-mini\",\"claude-3-5-sonnet\"]"
                        }
                    }
                },
                "description": "Request body for replacing a user's allowed model list"
            },
...
            "ResetPasswordResponse": {
                "type": "object",
                "properties": {
                    "temporaryPassword": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT access token in Authorization header: Bearer <token>",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Gateway client API key in Authorization header",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}